+++ opensrp-chw/build.gradle
@@ -261,7 +261,7 @@ android {
 }
 
 dependencies {
-    implementation('org.smartregister:opensrp-client-chw-core:1.1.37-SNAPSHOT@aar') {
+    implementation('org.smartregister:opensrp-client-chw-core:1.1.38-SNAPSHOT@aar') {
         transitive = true
         exclude group: 'com.android.support', module: 'appcompat-v7'
         exclude group: 'androidx.legacy', module: 'legacy-support-v4'
