+++ opensrp-chw/src/main/java/org/smartregister/chw/activity/FpRegisterActivity.java
@@ -7,8 +7,8 @@ import org.json.JSONException;
 import org.json.JSONObject;
 import org.smartregister.chw.R;
 import org.smartregister.chw.core.activity.CoreFpRegisterActivity;
-import org.smartregister.chw.dataloader.FPDataLoader;
-import org.smartregister.chw.form_data.NativeFormsDataBinder;
+import org.smartregister.chw.core.dataloader.FPDataLoader;
+import org.smartregister.chw.core.form_data.NativeFormsDataBinder;
 import org.smartregister.chw.fp.util.FamilyPlanningConstants;
 import org.smartregister.chw.fragment.FpRegisterFragment;
 import org.smartregister.family.util.JsonFormUtils;
