+++ opensrp-chw/src/main/java/org/smartregister/chw/dataloader/FamilyMemberDataLoader.java
@@ -11,8 +11,8 @@ import org.joda.time.Years;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
+import org.smartregister.chw.core.form_data.NativeFormsDataLoader;
 import org.smartregister.chw.core.utils.CoreConstants;
-import org.smartregister.chw.form_data.NativeFormsDataLoader;
 import org.smartregister.chw.util.Constants;
 import org.smartregister.chw.util.JsonFormUtils;
 import org.smartregister.chw.util.Utils;
