--- opensrp-chw/src/main/java/org/smartregister/chw/form_data/FormLoader.java
@@ -1,11 +1,0 @@
-package org.smartregister.chw.form_data;
-
-import android.content.Context;
-
-import org.json.JSONObject;
-
-public interface FormLoader {
-
-    JSONObject getJsonForm(Context context, String fileName) throws Exception;
-
-}
